# Conflicts:
#	.travis.yml
#	continuous_integration/test_script.sh
#	pystruct/learners/subgradient_latent_ssvm.py
#	pystruct/tests/test_learners/test_crammer_singer_svm.py
#	pystruct/tests/test_learners/test_graph_svm.py
#	setup.py

+++ pystruct/models/graph_crf.py
@@ -96,7 +96,7 @@ class GraphCRF(CRF):
         # try to set the size of joint_feature if possible
         if self.n_features is not None and self.n_states is not None:
             if self.directed:
-                self.size_joint_feature = (self.n_states * self.n_features +
+                self.size_joint_feature = int(self.n_states * self.n_features +
                                            self.n_states ** 2)
             else:
                 self.size_joint_feature = int(
